2. Organize project
